remove old api endpoint
